cylinder_sae: migrate to gamfit 0.1.123 thin wrapper
Route through gamfit.Cylinder as the basis descriptor; delete the
per-feature MLP encoder (fc1_w (F, D, H) — 210 GB at F=512 D=7168
H=14336 default) and replace with one shared nn.Linear → F·4 heads.
cylinder_sae_shared.py collapses to a 12-line alias.

Net: cylinder_sae.py 308→177 (−131); cylinder_sae_shared.py 177→12
(−165) = −296 LoC. Smoke F=128 K=16 3ep MPS on cogito-L40 val R²=0.846
(>0.7 sanity bar). 4/4 tests pass (also fixes pre-existing float64 dtype
mismatch in test_forward_shape).

gamfit gaps filed in memory: (1) Cylinder is a descriptor, not a
callable torch-grad basis evaluator; (2) sae_manifold_fit with
atom_topology='cylinder' rejects duchon_p/duchon_s kwargs and crashes
on the Duchon D2 dimension check.

+++ tests/test_cylinder_sae.py
@@ -46,7 +46,7 @@ def test_forward_shape() -> None:
     # TopK fires exactly top_k atoms per row
     assert torch.allclose(
         (out["amp"] > 0.5).float().sum(dim=1),
-        torch.full((B,), 4.0),
+        torch.full((B,), 4.0, dtype=torch.float32),
     )
     # basis size = (2H+1) * K_ell = 7 * 4 = 28
     assert model.M == 28
